% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roleQuery.R
\name{roleQueryUI}
\alias{roleQueryUI}
\title{查询角色UI}
\usage{
roleQueryUI(
  tabTitle = "查询角色",
  colTitles = c("操作区域", "报表区域"),
  widthRates = c(4, 8),
  func_left = roleQueryFunc_left,
  func_right = roleQueryFunc_right
)
}
\arguments{
\item{tabTitle}{标签名称}

\item{colTitles}{列名}

\item{widthRates}{比例}

\item{func_left}{函数}

\item{func_right}{函数}
}
\value{
返回值
}
\description{
查询角色UI
}
\examples{
roleQueryUI()
}
