% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roleRight.R
\name{roleRightUI}
\alias{roleRightUI}
\title{查询角色UI}
\usage{
roleRightUI(
  tabTitle = "角色授权",
  colTitles = c("操作区域", "系统对象区域"),
  widthRates = c(4, 8),
  func_left = roleRightFunc_left,
  func_right = roleRightFunc_right
)
}
\arguments{
\item{tabTitle}{标签名称}

\item{colTitles}{列名}

\item{widthRates}{比例}

\item{func_left}{函数}

\item{func_right}{函数}
}
\value{
返回值
}
\description{
查询角色UI
}
\examples{
roleRightUI()
}
