% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userAdd.R
\name{userAddUI}
\alias{userAddUI}
\title{用户新增UI}
\usage{
userAddUI(
  tabTitle = "创建用户",
  colTitles = c("操作区域", "报表区域"),
  widthRates = c(4, 8),
  func_left = userAddFunc_left,
  func_right = userAddFunc_right
)
}
\arguments{
\item{tabTitle}{标签名称}

\item{colTitles}{列名}

\item{widthRates}{比例}

\item{func_left}{函数}

\item{func_right}{函数}
}
\value{
返回值
}
\description{
用户新增UI
}
\examples{
userAddUI()
}
